# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwiftShip Logistics Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user password (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset

  # Service Stations
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorBody"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Riders
...
components:
  schemas:
...
    ResetPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, vendor, rider, staff]
        status:
          type: string
          enum: [active, inactive]
        stationId:
          type: ["integer", "null"]
        stationName:
          type: ["string", "null"]
        vendorId:
          type: ["integer", "null"]
        riderId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, vendor, rider, staff]
        stationId:
          type: ["integer", "null"]
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        stationId:
          type: ["integer", "null"]
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        vendorCode:
          type: string
        deliveryCharge:
          type: number
        status:
          type: string
          enum: [active, inactive]
        userId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - vendorCode
        - deliveryCharge
        - status
        - createdAt
    CreateVendorBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        vendorCode:
          type: string
        deliveryCharge:
          type: number
        userId:
          type: ["integer", "null"]
      required:
        - name
        - email
        - vendorCode
        - deliveryCharge
    UpdateVendorBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        vendorCode:
          type: string
        deliveryCharge:
          type: number
        status:
          type: string
        userId:
          type: ["integer", "null"]
...